@@ -3,6 +3,7 @@ package org.jabref.logic.database;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.jabref.logic.bibtex.comparator.BibtexStringComparator;
 import org.jabref.logic.importer.ImportFormatPreferences;
 import org.jabref.model.database.BibDatabase;
 import org.jabref.model.entry.BibEntry;
@@ -36,25 +37,51 @@ class DatabaseMergerTest {
     @Test
     void mergeAddsNonDuplicateEntries() {
         // Entries 1 and 2 are identical
-        BibEntry entry1 = new BibEntry()
+        BibEntry entry1 = new BibEntry(StandardEntryType.Article)
                 .withField(StandardField.AUTHOR, "Phillip Kaye and Michele Mosca")
                 .withField(StandardField.TITLE, "Quantum Networks for Generating Arbitrary Quantum States");
-        entry1.setType(StandardEntryType.Article);
-        BibEntry entry2 = new BibEntry()
+        BibEntry entry2 = new BibEntry(StandardEntryType.Article)
                 .withField(StandardField.AUTHOR, "Phillip Kaye and Michele Mosca")
                 .withField(StandardField.TITLE, "Quantum Networks for Generating Arbitrary Quantum States");
-        entry2.setType(StandardEntryType.Article);
-        BibEntry entry3 = new BibEntry()
+        BibEntry entry3 = new BibEntry(StandardEntryType.Article)
                 .withField(StandardField.AUTHOR, "Stephen Blaha")
                 .withField(StandardField.TITLE, "Quantum Computers and Quantum Computer Languages: Quantum Assembly Language and Quantum C Language");
-        entry3.setType(StandardEntryType.Article);
 
         BibDatabase database = new BibDatabase(List.of(entry1));
         BibDatabase other = new BibDatabase(List.of(entry2, entry3));
         new DatabaseMerger(importFormatPreferences.getKeywordSeparator()).merge(database, other);
 
-        assertEquals(2, database.getEntries().size());
         assertEquals(List.of(entry1, entry3), database.getEntries());
+    }
+
+    @Test
+    void mergeAddsWithDuplicateEntries() {
+        // Entries 1 and 2 are identical,  Entries 3 and 4 are identical
+        BibEntry entry1 = new BibEntry(StandardEntryType.Book)
+                .withField(StandardField.AUTHOR, "Joshua Bloch")
+                .withField(StandardField.TITLE, "Effective Java")
+                .withField(StandardField.ISBN, "0-201-31005-8");
+        BibEntry entry2 = new BibEntry(StandardEntryType.Book)
+                .withField(StandardField.AUTHOR, "Joshua J. Bloch")
+                .withField(StandardField.TITLE, "Effective Java:Programming Language Guide")
+                .withField(StandardField.ISBN, "0-201-31005-8")
+                .withField(StandardField.YEAR, "2001")
+                .withField(StandardField.EDITION, "1");
+        BibEntry entry3 = new BibEntry(StandardEntryType.Book)
+                .withField(StandardField.AUTHOR, "Joshua J. Bloch")
+                .withField(StandardField.TITLE, "Effective Java")
+                .withField(StandardField.ISBN, "978-0-321-35668-0")
+                .withField(StandardField.YEAR, "2008")
+                .withField(StandardField.EDITION, "2");
+        BibEntry entry4 = new BibEntry(StandardEntryType.Book)
+                .withField(StandardField.AUTHOR, "Joshua Bloch")
+                .withField(StandardField.TITLE, "Effective Java:Programming Language Guide")
+                .withField(StandardField.ISBN, "978-0-321-35668-0");
+
+        BibDatabase database = new BibDatabase(List.of(entry1, entry4));
+        BibDatabase other = new BibDatabase(List.of(entry2, entry3));
+        new DatabaseMerger(importFormatPreferences.getKeywordSeparator()).merge(database, other);
+        assertEquals(List.of(entry1, entry4), database.getEntries());
     }
 
     @Test
@@ -80,10 +107,10 @@ class DatabaseMergerTest {
         new DatabaseMerger(importFormatPreferences.getKeywordSeparator()).mergeStrings(target, source2);
         // Use string representation to compare since the id will not match
         List<String> resultStringsSorted = target.getStringValues()
-                                                 .stream()
-                                                 .map(BibtexString::toString)
-                                                 .sorted()
-                                                 .collect(Collectors.toList());
+                .stream()
+                .map(BibtexString::toString)
+                .sorted()
+                .collect(Collectors.toList());
 
         assertEquals(List.of(targetString.toString(), importedBibTeXString1.toString(),
                 importedBibTeXString2.toString()), resultStringsSorted);
@@ -106,14 +133,12 @@ class DatabaseMergerTest {
         source.addString(sourceString2);
 
         new DatabaseMerger(importFormatPreferences.getKeywordSeparator()).mergeStrings(target, source);
-        // Use string representation to compare since the id will not match
-        List<String> resultStringsSorted = target.getStringValues()
-                                                 .stream()
-                                                 .map(BibtexString::toString)
-                                                 .sorted()
-                                                 .collect(Collectors.toList());
+        List<BibtexString> resultStringsSorted = target.getStringValues()
+                .stream()
+                .sorted((s1, s2) -> new BibtexStringComparator(false).compare(s1, s2))
+                .collect(Collectors.toList());
 
-        assertEquals(List.of(targetString1.toString(), targetString2.toString()), resultStringsSorted);
+        assertEquals(List.of(targetString1, targetString2), resultStringsSorted);
     }
 
     @Test
@@ -162,6 +187,7 @@ class DatabaseMergerTest {
         assertEquals(targetRootGroup, target.getGroups().get());
         assertEquals(target.getGroups().get().getChildren().size(), 1);
         assertEquals(expectedImportedGroupNode, target.getGroups().get().getChildren().get(0));
+        assertEquals(expectedContentSelectors, target.getContentSelectorList());
     }
 
     static class TestGroup extends AbstractGroup {
