@@ -6,6 +6,7 @@ import com.tngtech.archunit.core.domain.JavaClasses;
 import com.tngtech.archunit.junit.AnalyzeClasses;
 import com.tngtech.archunit.junit.ArchIgnore;
 import com.tngtech.archunit.junit.ArchTest;
+import com.tngtech.archunit.library.GeneralCodingRules;
 
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.noClasses;
 import static com.tngtech.archunit.library.Architectures.layeredArchitecture;
@@ -20,6 +21,7 @@ class MainArchitectureTests {
     private static final String PACKAGE_ORG_JABREF_GUI = "org.jabref.gui..";
     private static final String PACKAGE_ORG_JABREF_LOGIC = "org.jabref.logic..";
     private static final String PACKAGE_ORG_JABREF_MODEL = "org.jabref.model..";
+    private static final String PACKAGE_ORG_JABREF_CLI = "org.jabref.cli..";
 
     @ArchTest
     public static void doNotUseApacheCommonsLang3(JavaClasses classes) {
@@ -92,7 +94,7 @@ class MainArchitectureTests {
                 .layer("Gui").definedBy(PACKAGE_ORG_JABREF_GUI)
                 .layer("Logic").definedBy(PACKAGE_ORG_JABREF_LOGIC)
                 .layer("Model").definedBy(PACKAGE_ORG_JABREF_MODEL)
-                .layer("Cli").definedBy("org.jabref.cli..")
+                .layer("Cli").definedBy(PACKAGE_ORG_JABREF_CLI)
                 .layer("Migrations").definedBy("org.jabref.migrations..") // TODO: Move to logic
                 .layer("Preferences").definedBy("org.jabref.preferences..")
                 .layer("Styletester").definedBy("org.jabref.styletester..")
@@ -133,6 +135,16 @@ class MainArchitectureTests {
         noClasses().that().resideInAPackage(PACKAGE_ORG_JABREF_LOGIC)
                    .should().dependOnClassesThat().resideInAPackage(PACKAGE_JAVAX_SWING)
                    .orShould().dependOnClassesThat().haveFullyQualifiedName(CLASS_ORG_JABREF_GLOBALS)
+                   .check(classes);
+    }
+
+    @ArchTest
+    public static void restrictStandardStreams(JavaClasses classes) {
+        noClasses().that().resideOutsideOfPackages(PACKAGE_ORG_JABREF_CLI)
+                   .and().resideOutsideOfPackages("org.jabref.gui.openoffice..") // Uses LibreOffice SDK
+                   .and().areNotAnnotatedWith(AllowedToUseStandardStreams.class)
+                   .should(GeneralCodingRules.ACCESS_STANDARD_STREAMS)
+                   .because("logging framework should be used instead or the class be marked explicitly as @AllowedToUseStandardStreams")
                    .check(classes);
     }
 }
