@@ -14,6 +14,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class FileAnnotationViewModelTest {
 
     private FileAnnotationViewModel annotationViewModel;
+    private FileAnnotationViewModel annotationViewModelWithoutFileAnnotation;
 
     @BeforeEach
     void setup() {
@@ -24,8 +25,9 @@ public class FileAnnotationViewModelTest {
 
         FileAnnotation linkedFileAnnotation = new FileAnnotation("John", LocalDateTime.now(), 3, content, FileAnnotationType.FREETEXT, Optional.empty());
         FileAnnotation annotation = new FileAnnotation("Jaroslav Kucha ˇr", LocalDateTime.parse("2017-07-20T10:11:30"), 1, marking, FileAnnotationType.HIGHLIGHT, Optional.of(linkedFileAnnotation));
-
+        FileAnnotation annotationWithoutFileAnnotation = new FileAnnotation("Jaroslav Kucha ˇr", LocalDateTime.parse("2017-07-20T10:11:30"), 1, marking, FileAnnotationType.HIGHLIGHT, Optional.empty());
         annotationViewModel = new FileAnnotationViewModel(annotation);
+        annotationViewModelWithoutFileAnnotation = new FileAnnotationViewModel(annotationWithoutFileAnnotation);
     }
 
     @Test
@@ -49,11 +51,25 @@ public class FileAnnotationViewModelTest {
     }
 
     @Test
+    public void retrieveCorrectContentWithoutLinkedFileAnnotation() {
+        String expectedMarking = String.format("This is paragraph 1.%n" +
+                "This is paragraph 2, and it crosses several lines, now you can see next paragraph:%n"
+                + "This is paragraph 3.");
+
+        assertEquals(expectedMarking, annotationViewModelWithoutFileAnnotation.getContent());
+    }
+
+    @Test
     public void removeOnlyLineBreaksNotPrecededByPeriodOrColon() {
         String expectedMarking = String.format("This is paragraph 1.%n" +
                 "This is paragraph 2, and it crosses several lines, now you can see next paragraph:%n"
                 + "This is paragraph 3.");
 
         assertEquals(expectedMarking, annotationViewModel.getMarking());
+    }
+
+    @Test
+    public void retrieveCorrectMarkingWithoutLinkedFileAnnotation() {
+        assertEquals("", annotationViewModelWithoutFileAnnotation.getMarking());
     }
 }
